+++ api/objects/user.php
@@ -42,13 +42,16 @@ class User{
         $new_user->totp_secret = $this->totp_secret;
         $new_user->totp_enabled = $this->totp_enabled;
         $new_user->password = $password_hash;
-        $new_user->group = "users";
+        $new_user->appGroup = "users";
+        try {
         $new_user->save();
         if($new_user->id()){
             return true;
         }
-    
-        return false;
+        } catch (exception $e) {
+            $t = $e->getMessage();
+            return false;
+        }         
     }
         
     // enable totp for user